Fix regen to work with older versions of sympy
The problem was the sympy.Derivative class, specifically its args
attribute.  In older versions, the args of a second derivative would
be

    (u(x,y), x, x)

and in the newer versions of sympy, it would be

    (u(x,y), (x,2))

which is a different number of elements in the tuple.  The new logic
works with both return types.