Added a SummaryReport notebook, and two static versions of it
The static versions are SummaryReport.html and SummaryReport.pdf.  To
create the HTML version, I ran

    $ jupyter nbconvert SummaryReport.ipynb

To generate the PDF version, I opened the HTML version in a browser,
chose to Print, and from the dialogue box, chose to save it as a PDF.

There are more sophsticted ways to generate a LaTeX-quality report
from a Jupyter Notebook, but I didn't have the time to learn enough of
the details.