Added jupyter_contrib_nbextensions to the list of Prerequisites
I did this as I was testing jupyter_contrib_nbextensions as a package
to help me convert a notebook to a PDF report.  Ultimately, the
package proved too complicated to learn in the time allotted to the
project, but it certainly looks cool.  So this package is optional,
but if the user installs it, they will see that equations in markdown
cells get numbered!